TECH-2322 expose es6 modules: extract filter menu classes into files
